Academic course data for 98,000 undergraduates

A dataset of course attributes of 97,064 undergraduates matriculating in US universities from from 1987 to 2016, sampled from the MIDFIELD database.

midfieldcourses

Format

A data frame (tibble) with 3.5 M observations and 12 variables, occupying 348 Mb of memory. Each observation is one course in one term for one student. The variables are:

id

Unique, anonymized MIDFIELD student identifier.

institution

Anonymized institution name.

term_course

The term the course was attempted

course

Course name

abbrev

Course alpha identifier (e.g. ENGR, MATH, ENGL)

number

Course numeric identifier (e.g. 101, 3429)

section

Course section identifier

hours_course

The number of hours awarded upon successful completion of the course

type

The means by which instruction is predominately delivered for this course section

pass_fail

Was the course offered pass/fail?

grade

The grade awarded for the course

faculty_rank

Academic rank of the person teaching the course

Source

Data provided by the MIDFIELD project: https://engineering.purdue.edu/MIDFIELD

Examples

# NOT RUN {
tibble::glimpse(midfieldcourses)
# }