Demographic data for 98,000 undergraduates

Data frame of matriculation attributes with 97,640 observations and 15 demographic variables keyed by student ID. Each observation is a unique student. A stratified sample of the MIDFIELD database.

midfieldstudents

Format

data.table with 97,064 observations and 15 variables, occupying 19 MB of memory. Each observation is a unique student. The variables are:

id

character, unique anonymized MIDFIELD student identifier

institution

character, anonymized institution name

cip6

character, 6-digit CIP code of instructional program at matriculation

term_enter

numeric, academic term the student first entered the institution, format YYYYT

transfer

character, did the student transfer into the institution? (Y = Yes or N = No)

hours_transfer

numeric, transfer hours accepted at the institution

sex

character, sex (M = Male, F = Female) as self-reported by the student

race

character, race/ethnicity as self-reported by the student and as reported to IPEDS

age

numeric, age at matriculation as self-reported by the student

us_citizen

character, is the student a US citizen? (Y = Yes or N = No)

home_zip

character, five or nine digit home ZIP code at time of entry

high_school

character, high school code for the last high school attended before entering a post-secondary institution

sat_math

numeric, SAT mathematics test score

sat_verbal

numeric, SAT reading test score

act_comp

numeric, ACT composite test score

Source

Data provided by the MIDFIELD project: https://engineering.purdue.edu/MIDFIELD

Examples

if (FALSE) {
library(data.table)
midfieldstudents
}