Course data sample

Table of course-related attributes of approximately 98,000 undergraduates enrolled in courses at their institutions. Each observation is one course in one term for one student keyed by student ID and a combination of abbreviation, number, and section.

Usage

data(course)

Format

A data.table with approximately 3.4 M observations and 12 variables occupying 340 MB of memory:

mcid

Character, anonymized student identifier

institution

Character, anonymized institution name, e.g., Institution A, Institution B, etc.

term

Character, academic year and term, format YYYYT

course

Character, course name, e.g., Chemistry, College Algebra, US History, etc.

abbrev

Character, course alpha identifier, e.g. ENGR, MATH, ENGL, composite key (see id)

number

Character, course numeric identifier, e.g. 101, 3429, composite key (see id)

section

Character, course section identifier, composite key (see id)

type

Character, predominant delivery method for this section, e.g., Blended, Honors, Lecture, Seminar, etc.

faculty_rank

Character, academic rank of the person teaching the course, e.g., Assistant Professor, Associate professor, Graduate Assistant, Visiting, etc.

hours_course

Numeric, number of credit-hours for successful course completion

pass_fail

Character, whether the course was offered on a pass/fail basis, No, Yes, or missing

grade

Character, course grade, e.g., A+, A, A-, B+, I, NG, etc.

Source

Data provided by the MIDFIELD project: https://midfield.online/

Details

The data in midfielddata are a proportionate stratified random sample of 12 institutions in the MIDFIELD research database. However, these are practice data, not research data---suitable for learning about student-record analysis, but not for drawing inferences about student performance.

Students in the course table have records in the student table only if they are admitted as degree-seeking undergraduates and in the degree table only if they complete a program.

Terms are encoded YYYYT, where YYYY is the year of the Fall term that starts the academic year and T = 1 (Fall), 2 (Winter, quarter systems only), 3 (Spring), and 4 (Summer) or 5 and 6 (Summer sessions 1 and 2). For example, for the 1995--96 academic year, Fall '95 is encoded 19951, Spring '96 is encoded 19953, and the first summer '96 term is encoded 19954. The MIDFIELD research database also includes terms of one month duration, encoded with letters A, B, C, etc., though none are included in this sample.