midfielddata: MIDFIELD Data Sample

This data package contains anonymized student-level records for 98,000 undergraduates from three US institutions from 1988 through 2018 organized in four data tables keyed by student ID. Student-level data refers to information collected by undergraduate institutions on individual students including demographics, programs and academic standing by term, individual courses, grades, and degrees. These data are a proportionate stratified sample of the MIDFIELD database, but are not suitable for drawing inferences about program attributes or student experiences—‘midfielddata’ provides practice data, not research data.

See also

Author

Maintainer: Richard Layton graphdoctor@gmail.com [copyright holder]

Authors:

Russell Long ralong@purdue.edu [copyright holder, data manager]

Matthew Ohland ohland@purdue.edu [copyright holder]

Marisa Orr marisak@clemson.edu [copyright holder]

Susan Lord slord@sandiego.edu [copyright holder]

Other contributors:

US National Science Foundation [funder]