Demographic data for 98,000 undergraduates

A dataset of demographic attributes of 97,640 undergraduates matriculating in US universities from from 1987 to 2016, sampled from the MIDFIELD database.

midfieldstudents

Format

A data frame (tibble) with 97,064 observations and 15 variables, occupying 19 MB of memory. Each observation is a unique student. The variables are:

id

Unique, anonymized MIDFIELD student identifier.

institution

Anonymized institution name.

cip6

Six-digit CIP code of the program into which the student matriculated.

term_enter

Academic year and term the student first entered the institution.

transfer

Did the student transfer into the institution? (Y = Yes or N = No).

hours_transfer

Transfer hours accepted at your institution.

sex

Sex (M = Male, F = Female) as self-reported by the student.

race

Ethnicity as self-reported by the student and as reported to IPEDS.

age

Age at matriculation as self-reported by the student.

us_citizen

Is the student a US citizen? (Y = Yes or N = No).

home_zip

Five or nine digit home ZIP code at time of entry.

high_school

High school code for the last high school attended before entering a post-secondary institution

sat_math

SAT mathematics test score, used as part of the admissions process.

sat_verbal

SAT reading test score, used as part of the admissions process.

act_comp

ACT composite test score, used as part of the admissions process.

Source

Data provided by the MIDFIELD project: https://engineering.purdue.edu/MIDFIELD

Examples

# NOT RUN {
tibble::glimpse(midfieldstudents)
# }