Student data sample

Table of attributes of approximately 98,000 students upon being admitted as degree-seeking undergraduates. Each observation is a unique student keyed by student ID.

student

Format

A data.table with approximately 98,000 observations and 13 variables occupying 19 MB of memory:

mcid

character, anonymized student identifier

institution

character, anonymized institution name, e.g., Institution A, Institution B, etc.

transfer

character, stating whether the student is a First-Time in College students or a First-Time Transfer student

hours_transfer

numeric, transfer hours accepted at the institution

race

character, race/ethnicity as self-reported by the student, e.g., Asian, Black, Hispanic/LatinX, etc.

sex

character, sex as self-reported by the student, values are Female, Male, and Unknown

age

numeric, age at admission as self-reported by the student

us_citizen

character, is the student a US citizen? (Yes or No)

home_zip

character, five or nine digit home ZIP code at admission

high_school

character, high school code for the last high school attended before entering a post-secondary institution

sat_math

numeric, SAT mathematics test score

sat_verbal

numeric, SAT reading test score

act_comp

numeric, ACT composite test score

Source

Data provided by the MIDFIELD project: https://engineering.purdue.edu/MIDFIELD

Details

The data in midfielddata are a proportionate stratified random sample of 12 institutions in the MIDFIELD research database. However, these are practice data, not research data---suitable for learning about student-record analysis, but not for drawing inferences about student performance.

Students in the student table have records in the term table only if they complete a term; in the course table if they enroll in a course; and in the degree table only if they complete a program.

Examples

if (FALSE) {
library(data.table)
student
}