Student data sample

Demographic information at matriculation for approximately 98,000 degree-seeking undergraduate students. One record per student, keyed by student ID.

Usage

data(student, package = "midfielddata")

Format

A data.frame and data.table with 13 variables and 97,555 observations of unique students occupying 18 MB of memory:

mcid

Character, anonymized student identifier, e.g., "MCID3111142225".

institution

Character, anonymized institution name, e.g., "Institution B".

transfer

Character, transfer status, possible values are "First-Time in College", "First-Time Transfer".

hours_transfer

Numeric, number of credit hours transferred (or NA).

race

Character, race/ethnicity, e.g., "Asian", "Black", "International", "Latine", etc.

sex

Character, sex, possible values are "Female", "Male", "Unknown".

age_desc

Character, age group, possible values are "25 and Older", "Under 25".

us_citizen

Character, US citizenship, possible values are "No", "Yes".

home_zip

Character, home ZIP code (or NA), e.g., "02056", "20170", "51301", "80129", etc.

high_school

Character, code for the last high school attended before admission (or NA), e.g., "060075", "210512", "431800", "502195", etc.

sat_math

Numeric, SAT mathematics test score (or NA).

sat_verbal

Numeric, SAT reading test score (or NA).

act_comp

Numeric, ACT composite test score (or NA).

Source

Data provided by the MIDFIELD project: https://midfield.online/

Details

The data in 'midfielddata' are practice data, suitable for learning to work with Student Unit Records (SURs) generally. Unlike the MIDFIELD research database, the data tables in 'midfielddata' are not research data; they are not suitable for drawing inferences about program attributes or student experiences.